+++ apollo-router/src/plugins/telemetry/config_new/instruments.rs
@@ -30,9 +30,9 @@ use crate::plugins::telemetry::config_new::conditions::Condition;
 use crate::plugins::telemetry::config_new::cost::CostInstruments;
 use crate::plugins::telemetry::config_new::cost::CostInstrumentsConfig;
 use crate::plugins::telemetry::config_new::extendable::Extendable;
-use crate::plugins::telemetry::config_new::graphql::GraphQLAttributes;
+use crate::plugins::telemetry::config_new::graphql::attributes::GraphQLAttributes;
+use crate::plugins::telemetry::config_new::graphql::selectors::GraphQLSelector;
 use crate::plugins::telemetry::config_new::graphql::GraphQLInstrumentsConfig;
-use crate::plugins::telemetry::config_new::graphql::GraphQLSelector;
 use crate::plugins::telemetry::config_new::selectors::RouterSelector;
 use crate::plugins::telemetry::config_new::selectors::SubgraphSelector;
 use crate::plugins::telemetry::config_new::selectors::SupergraphSelector;
